Added CPU implementation.
Also adds optimizations to automatically replace the CPU Ops w/GPU ones
when possible. TODO: Performance of the CRF inference is slightly slower
when the CPU Ops are automatically replaced w/GPU ones as opposed to
directly using GPU Ops. It's not clear why this is the case but the
extra time is spent in GpuHashTable within Scan.

Closes #1.